requirements.txt

django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
google-generativeai==0.3.1
faiss-cpu==1.7.4
pypdf==3.17.1
python-dotenv==1.0.0
